+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to
 
 
 ## [Unreleased]
+
+## [0.1.4] — 2019-08-10
 ### Added
 - `read-all-bytes` method.
 - `create-temp-file` method.
@@ -57,4 +59,5 @@ and this project adheres to
 [0.1.1]: https://github.com/logicblocks/pathological/compare/0.1.0...0.1.1
 [0.1.2]: https://github.com/logicblocks/pathological/compare/0.1.1...0.1.2
 [0.1.3]: https://github.com/logicblocks/pathological/compare/0.1.2...0.1.3
-[Unreleased]: https://github.com/logicblocks/pathological/compare/0.1.3...HEAD
+[0.1.4]: https://github.com/logicblocks/pathological/compare/0.1.3...0.1.4
+[Unreleased]: https://github.com/logicblocks/pathological/compare/0.1.4...HEAD